// frontend/src/client/models/ChatMessage.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ChatMessage = {
    content: string;
    role: ChatMessage.role;
};

export namespace ChatMessage {

    export enum role {
        USER = 'user',
        SYSTEM = 'system',
        ASSISTANT = 'assistant',
    }


}